# api-definition.inc.yml
...
swagger: '2.0'
info:
  description: Server to manage and analyze Threat Models
  version: 1.0.0
  title: Threat Model Assessment App
...
paths:
...
  /model/save:
    post:
      tags:
        - model
      summary: Save dfd models
      description: Save dfd models from the frontend
      operationId: saveDFDModel
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: DFD Model
          schema:
            $ref: '#/definitions/BaseDFDModel'
        - in: header
          name: Authorization
          description: Authorization header
          required: true
          type: string
      responses:
        '201':
          description: Saved DFD model
        '400':
          description: Incorrect message format
...
  /model/threat-update:
    post:
      tags:
        - model
      summary: Update threat properties
      description: Update threat properties to set applicable state and priority
      operationId: updateSTRIDEThreatData
      produces:
        - application/json
      parameters:
        - in: header
          name: Authorization
          description: Authorization header
          required: true
          type: string
        - in: body
          name: body
          description: Threat Data
          schema:
            $ref: '#/definitions/AppliedSTRIDEThreat'
      responses:
        '200':
          description: Updated threat successfully
        '401':
          description: Unauthorized
  /model/analyze:
    post:
      tags:
        - model
      summary: Analyze a new threat model
      description: Return the analysis of a threat model
      operationId: analyzeThreatModel
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Threat Model
          schema:
            $ref: '#/definitions/AnalysisDFDModel'
        - in: header
          name: Authorization
          description: Authorization header
          required: true
          type: string
      responses:
        '200':
          description: Analysis successful
          schema:
            $ref: '#/definitions/AnalysisResult'
        '401':
          description: Unauthorized
  '/model/analyze/{modelID}':
    get:
      tags:
        - model
      summary: Analyze an existing threat model
      description: >-
        Return the analyis of a threat model that is already stored in the
        database
      operationId: analyzeStoredThreatModel
      produces:
        - application/json
      parameters:
        - in: path
          name: modelID
          description: ID of the saved model
          required: true
          type: string
        - in: header
          name: Authorization
          description: Authorization header
          required: true
          type: string
      responses:
        '200':
          description: Analysis successful
          schema:
            $ref: '#/definitions/AnalysisResult'
        '401':
          description: Unauthorized
  '/model/{modelID}':
    get:
      tags:
        - model
      summary: Load a specific model
      description: Loads a model previously saved by a user
      operationId: loadModel
      produces:
        - application/json
      parameters:
        - in: path
          name: modelID
          description: ID of the saved model
          required: true
          type: string
        - in: header
          name: Authorization
          description: Authorization header
          required: true
          type: string
      responses:
        '200':
          description: Found model
          schema:
            $ref: '#/definitions/StoredDFDModel'
        '400':
          description: Bad request
        '404':
          description: No model found
...
definitions:
...
  BaseModelStorageData:
    type: object
    properties:
      name:
        type: string
      projectTitle:
        type: string
      summary:
        type: string
      screenshot:
        type: string
      projectID:
        type: string
...
  CPE:
    type: object
    properties:
      id:
        type: integer
        format: int32
      title:
        type: string
      references:
        type: array
        items:
          $ref: '#/definitions/CPEReference'
      cpeName:
        type: string
      cpe23Name:
        type: string
  CPEReference:
    type: object
    properties:
      referenceType:
        type: string
      referenceContent:
        type: string
  CVE:
    type: object
    description: Contains all data available for a vulnerability
    properties:
      cveID:
        type: string
      assigner:
        type: string
      dataFormat:
        type: string
      dataVersion:
        type: string
      publishedDate:
        type: string
      lastModifiedDate:
        type: string
      affects:
        type: array
        items:
          $ref: '#/definitions/VendorData'
      problemTypes:
        type: array
        items:
          $ref: '#/definitions/ProblemType'
      references:
        type: array
        items:
          $ref: '#/definitions/Reference'
      descriptions:
        type: array
        items:
          $ref: '#/definitions/VulnerabilityDescription'
      vulnerableConfig:
        $ref: '#/definitions/VulnerableConfiguration'
      cvss:
        $ref: '#/definitions/CVSS'
  VendorData:
    type: object
    description: Contains the vendor name and a list of affected products
    properties:
      vendorName:
        type: string
      products:
        type: array
        items:
          $ref: '#/definitions/Product'
...
  ProblemType:
    type: object
    description: Contains a list of problem type meta data
    properties:
      descriptions:
        type: array
        items:
          $ref: '#/definitions/ProblemTypeDescription'
...
  Reference:
    type: object
    description: >-
      Contains further information about a vulnerability and sometimes
      mitigation suggestions
    properties:
      url:
        type: string
      name:
        type: string
      refSource:
        type: string
      tags:
        type: array
        items:
          type: string
  VulnerabilityDescription:
    type: object
    description: Contains the description of a vulnerability
    properties:
      language:
        type: string
      content:
        type: string
  VulnerableConfiguration:
    type: object
    description: Contains the data specifying the affected software configurations by CPE
    properties:
      cveDataVersion:
        type: string
      configs:
        type: array
        items:
          $ref: '#/definitions/Config'
...
  CVSS:
    type: object
    description: >-
      Contains both the CVSS v3 and CVSS v2 metric describing a vulnerabilities
      impact
    properties:
      cvssv3Metric:
        $ref: '#/definitions/CVSSV3Metric'
      cvssv2Metric:
        $ref: '#/definitions/CVSSV2Metric'
...
  BaseDFDModel:
    type: object
    properties:
      interactors:
        type: array
        items:
          $ref: '#/definitions/Interactor'
      dataFlows:
        type: array
        items:
          $ref: '#/definitions/DataFlow'
      dataStores:
        type: array
        items:
          $ref: '#/definitions/DataStore'
      processes:
        type: array
        items:
          $ref: '#/definitions/Process'
      trustBoundaries:
        type: array
        items:
          $ref: '#/definitions/TrustBoundary'
      storageData:
        $ref: '#/definitions/BaseModelStorageData'
  StoredDFDModel:
    type: object
    properties:
      modelID:
        type: string
      interactors:
        type: array
        items:
          $ref: '#/definitions/Interactor'
      dataFlows:
        type: array
        items:
          $ref: '#/definitions/DataFlow'
      dataStores:
        type: array
        items:
          $ref: '#/definitions/DataStore'
      processes:
        type: array
        items:
          $ref: '#/definitions/Process'
      trustBoundaries:
        type: array
        items:
          $ref: '#/definitions/TrustBoundary'
  AnalysisDFDModel:
    type: object
    properties:
      modelID:
        type: string
      interactors:
        type: array
        items:
          $ref: '#/definitions/AnalysisInteractor'
      dataFlows:
        type: array
        items:
          $ref: '#/definitions/AnalysisDataFlow'
      dataStores:
        type: array
        items:
          $ref: '#/definitions/AnalysisDataStore'
      processes:
        type: array
        items:
          $ref: '#/definitions/AnalysisProcess'
      boundaries:
        type: array
        items:
          $ref: '#/definitions/AnalysisBoundary'
  Interactor:
    allOf:
      - $ref: '#/definitions/AnalysisInteractor'
      - type: object
        properties:
          ip:
            type: string
          coordinates:
            $ref: '#/definitions/PointCoordinates'
          genericType:
            $ref: '#/definitions/ElementType'
          cpe:
            $ref: '#/definitions/CPE'
          connectedToDataFlow:
            type: boolean
  Process:
    allOf:
      - $ref: '#/definitions/AnalysisProcess'
      - type: object
        properties:
          ip:
            type: string
          coordinates:
            $ref: '#/definitions/PointCoordinates'
          genericType:
            $ref: '#/definitions/ElementType'
          cpe:
            $ref: '#/definitions/CPE'
          connectedToDataFlow:
            type: boolean
  DataStore:
    allOf:
      - $ref: '#/definitions/AnalysisDataStore'
      - type: object
        properties:
          ip:
            type: string
          coordinates:
            $ref: '#/definitions/PointCoordinates'
          genericType:
            $ref: '#/definitions/ElementType'
          cpe:
            $ref: '#/definitions/CPE'
          connectedToDataFlow:
            type: boolean
  DataFlow:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      connectedElements:
        $ref: '#/definitions/DataFlowConnectionInfo'
      position:
        $ref: '#/definitions/DataFlowDrawingPosition'
      type:
        $ref: '#/definitions/DataFlowType'
      options:
        $ref: '#/definitions/DataFlowOptions'
      genericType:
        $ref: '#/definitions/ElementType'
      cpe:
        $ref: '#/definitions/CPE'
  TrustBoundary:
    allOf:
      - $ref: '#/definitions/AnalysisBoundary'
      - type: object
        properties:
          coordinates:
            $ref: '#/definitions/PointCoordinates'
          width:
            type: number
          height:
            type: number
          genericType:
            $ref: '#/definitions/ElementType'
  AnalysisInteractor:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/InteractorType'
      options:
        $ref: '#/definitions/InteractorOptions'
      cpe:
        type: string
  AnalysisProcess:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/ProcessType'
      options:
        $ref: '#/definitions/ProcessOptions'
      cpe:
        type: string
  AnalysisDataStore:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/DataStoreType'
      options:
        $ref: '#/definitions/DataStoreOptions'
      cpe:
        type: string
  AnalysisDataFlow:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      startElement:
        $ref: '#/definitions/ConnectedElement'
      endElement:
        $ref: '#/definitions/ConnectedElement'
      type:
        $ref: '#/definitions/DataFlowType'
      options:
        $ref: '#/definitions/DataFlowOptions'
      cpe:
        type: string
  AnalysisBoundary:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/BoundaryType'
      options:
        $ref: '#/definitions/BoundaryOptions'
  AnalysisResult:
    type: object
    properties:
      modelID:
        type: string
      securityThreats:
        type: array
        items:
          $ref: '#/definitions/AppliedSTRIDEThreat'
      privacyThreats:
        type: array
        items:
          $ref: '#/definitions/AppliedSTRIDEThreat'
      vulnerabilities:
        type: array
        items:
          $ref: '#/definitions/CVE'
  ElementType:
    type: string
    enum:
      - process
      - interactor
      - datastore
      - dataflow
      - trustboundary
  InteractorType:
    type: string
    enum:
      - Generic Interactor
      - Web Application
      - Mobile Application
      - Human
      - Browser
      - IoT Gateway
      - Sensor
      - Operating System
      - External System
      - External API
  ProcessType:
    type: string
    enum:
      - Generic Process
      - Web Server
      - Virtual Machine
      - Thick Client
      - Thin Client
      - Thread
      - Application
      - Web Service
  DataStoreType:
    type: string
    enum:
      - Generic Data Store
      - SQL Database
      - NoSQL Database
      - Cookies
      - Session Storage
      - Cloud Storage
      - Device
      - File
      - Cache
  DataFlowType:
    type: string
    enum:
      - Generic Data Flow
      - HTTP
      - HTTPS
      - TCP/IP
      - UDP
      - IPSec
      - MQTT
      - FTP
      - SSH
      - SMTP
      - Custom
      - Binary
  BoundaryType:
    type: string
    enum:
      - Generic Trust Boundary
      - Internal Network
      - External Network
      - Physical Boundary
      - Local System
  InteractorOptions:
    type: object
    properties:
      trustLevel:
        $ref: '#/definitions/ScaleSelection'
      privilegeLevel:
        $ref: '#/definitions/ScaleSelection'
      sanitizesInput:
        $ref: '#/definitions/GenericSelection'
      employee:
        $ref: '#/definitions/GenericSelection'
      accessible:
        $ref: '#/definitions/GenericSelection'
      appType:
        $ref: '#/definitions/AppType'
  ProcessOptions:
    type: object
    properties:
      trustLevel:
        $ref: '#/definitions/ScaleSelection'
      privilegeLevel:
        $ref: '#/definitions/ScaleSelection'
      managed:
        $ref: '#/definitions/GenericSelection'
      dataHandling:
        $ref: '#/definitions/DataHandling'
      inputOrigin:
        $ref: '#/definitions/InputOrigin'
      sanitizesInput:
        $ref: '#/definitions/GenericSelection'
      validatesInput:
        $ref: '#/definitions/GenericSelection'
      requiresAuthentication:
        $ref: '#/definitions/GenericSelection'
      authenticatesItself:
        $ref: '#/definitions/GenericSelection'
      ddosProtection:
        $ref: '#/definitions/GenericSelection'
      customCommunicationProtocol:
        $ref: '#/definitions/GenericSelection'
  DataStoreOptions:
    type: object
    properties:
      trustLevel:
        $ref: '#/definitions/ScaleSelection'
      privilegeLevel:
        $ref: '#/definitions/ScaleSelection'
      storesSensitiveData:
        $ref: '#/definitions/GenericSelection'
      storesPersonalData:
        $ref: '#/definitions/GenericSelection'
      encrypted:
        $ref: '#/definitions/GenericSelection'
      removeableStorage:
        $ref: '#/definitions/GenericSelection'
      backup:
        $ref: '#/definitions/GenericSelection'
      shared:
        $ref: '#/definitions/GenericSelection'
      shellEnabled:
        $ref: '#/definitions/GenericSelection'
      storesLogs:
        $ref: '#/definitions/GenericSelection'
  DataFlowOptions:
    type: object
    properties:
      transfersSensitiveData:
        $ref: '#/definitions/GenericSelection'
      networkType:
        $ref: '#/definitions/NetworkType'
      payload:
        $ref: '#/definitions/PayloadType'
      trustedNetwork:
        $ref: '#/definitions/GenericSelection'
  BoundaryOptions:
    type: object
    properties:
      authenticatesItself:
        $ref: '#/definitions/GenericSelection'
      destinationAuthenticated:
        $ref: '#/definitions/GenericSelection'
      human:
        $ref: '#/definitions/GenericSelection'
  GenericSelection:
    type: string
    enum:
      - true
      - false
      - notSelected
  ScaleSelection:
    type: string
    enum:
      - Low
      - Medium
      - High
      - Not Selected
  AppType:
    type: string
    enum:
      - Native
      - Cross-Platform
      - Hybrid
      - Progressive Web App
      - Not Selected
  DataHandling:
    type: string
    enum:
      - Personal Data
      - Sensitive Data
      - Both
      - Other
      - Not Selected
  InputOrigin:
    type: string
    enum:
      - Anywhere
      - Interal Network
      - Local System
      - Not Selected
  NetworkType:
    type: string
    enum:
      - Wire
      - Wi-Fi
      - Bluetooth
      - Mobile Network
      - Other
      - Not Selected
  PayloadType:
    type: string
    enum:
      - JSON
      - XML
      - Query String
      - SOAP
      - Other
      - Not Selected
  PointCoordinates:
    type: object
    properties:
      x:
        type: number
        format: float
      'y':
        type: number
        format: float
  ConnectedElement:
    type: object
    properties:
      id:
        type: string
      type:
        $ref: '#/definitions/ElementType'
  PositionedConnectedElement:
    allOf:
      - $ref: '#/definitions/ConnectedElement'
      - type: object
        properties:
          coordinates:
            $ref: '#/definitions/PointCoordinates'
  DataFlowConnectionInfo:
    type: object
    properties:
      startElement:
        $ref: '#/definitions/PositionedConnectedElement'
      endElement:
        $ref: '#/definitions/PositionedConnectedElement'
  DataFlowDrawingPosition:
    type: string
    enum:
      - top
      - normal
      - bottom
...